Putting WildcardTree in akka.util and making it a bit more generic
